Adjust for single library header
